Fixing includes after restructure
